+++ sources/framework/sandbox/sandbox.js
@@ -31,7 +31,7 @@
             var me = this;
 
             me._core = core;
-            this._map = Oskari.clazz.create('Oskari.mapframework.domain.Map');
+            this._map = Oskari.clazz.create('Oskari.mapframework.domain.Map', this);
 
             /*
              * All registered listeners in map key: event name value: